+++ fabric/src/main/java/io/github/sakurawald/fuji/core/auxiliary/LogUtil.java
@@ -5,10 +5,8 @@ import io.github.sakurawald.fuji.core.config.Configs;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.loader.api.FabricLoader;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.apache.logging.log4j.core.config.Configurator;
 import org.jetbrains.annotations.NotNull;
 
 public class LogUtil {
@@ -16,19 +14,7 @@ public class LogUtil {
     private static final @NotNull Logger MOD_LOGGER = makeLogger(StringUtils.capitalize(Fuji.MOD_ID));
 
     private static @NotNull Logger makeLogger(String name) {
-        Logger logger = LogManager.getLogger(name);
-        try {
-            configureLogger(logger);
-        } catch (Exception e) {
-            return logger;
-        }
-        return logger;
-    }
-
-    private static void configureLogger(Logger logger) {
-        // You can see the `debug` logs in `logs/debug.txt` file
-        String levelDefinedInEnv = System.getProperty("%s.level".formatted(Fuji.MOD_ID));
-        Configurator.setLevel(logger, Level.getLevel(levelDefinedInEnv));
+        return LogManager.getLogger(name);
     }
 
     private static final boolean isConsoleSupportAnsiColor = isConsoleSupportAnsiColor();